+++ src/main/java/io/github/astrapi69/xml/xpath/NodeExtensions.java
@@ -26,6 +26,8 @@ package io.github.astrapi69.xml.xpath;
 
 import java.io.StringReader;
 import java.io.StringWriter;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 
 import javax.xml.parsers.ParserConfigurationException;
@@ -47,12 +49,77 @@ import io.github.astrapi69.xml.transform.TransformerFactoryInitializer;
  */
 public class NodeExtensions
 {
-
+	/**
+	 * Resolves the node from the given index and returns it as {@link String} object
+	 * 
+	 * @param nodeList
+	 *            the node list
+	 * @param index
+	 *            the index from the node
+	 * @return the node from the given index and returns it as {@link String} object
+	 * @throws TransformerException
+	 *             is thrown if an unrecoverable error occurs during the course of the
+	 *             transformation
+	 */
 	public static String toString(NodeList nodeList, int index) throws TransformerException
 	{
 		return toString(nodeList.item(index));
 	}
 
+	/**
+	 * Returns the given {@link NodeList} object as {@link String} object
+	 *
+	 * @param nodeList
+	 *            the node list
+	 * @return the given {@link NodeList} object as {@link String} object
+	 * @throws TransformerException
+	 *             is thrown if an unrecoverable error occurs during the course of the
+	 *             transformation
+	 */
+	public static String toString(NodeList nodeList) throws TransformerException
+	{
+		int length = nodeList.getLength();
+		StringBuilder stringBuilder = new StringBuilder();
+		for (int i = 0; i < length; i++)
+		{
+			Node item = nodeList.item(i);
+			stringBuilder.append(toString(item));
+		}
+		return stringBuilder.toString();
+	}
+
+	/**
+	 * Returns the content of the values from the given {@link NodeList} object as {@link List} of
+	 * {@link String} objects
+	 * 
+	 * @param nodeList
+	 *            the node list
+	 * @return the content of the values from the given {@link NodeList} object as {@link List} of
+	 *         {@link String} objects
+	 */
+	public static List<String> toValueList(NodeList nodeList)
+	{
+		int length = nodeList.getLength();
+		List<String> result = new ArrayList<>(length);
+		for (int i = 0; i < length; i++)
+		{
+			Node item = nodeList.item(i);
+			String stringValue = toStringValue(item);
+			result.add(stringValue);
+		}
+		return result;
+	}
+
+	/**
+	 * Returns a xml representation as {@link String} object of the given {@link Node} object
+	 * 
+	 * @param node
+	 *            the {@link Node} object
+	 * @return a xml representation as {@link String} object of the given {@link Node} object
+	 * @throws TransformerException
+	 *             is thrown if an unrecoverable error occurs during the course of the
+	 *             transformation
+	 */
 	public static String toString(Node node) throws TransformerException
 	{
 		StringWriter stringWriter = new StringWriter();
@@ -61,6 +128,30 @@ public class NodeExtensions
 		return stringWriter.toString();
 	}
 
+	/**
+	 * Returns the text content as {@link String} object of the given {@link Node} object
+	 * 
+	 * @param node
+	 *            the {@link Node} object
+	 * @return the text content as {@link String} object of the given {@link Node} object
+	 */
+	public static String toStringValue(Node node)
+	{
+		return node.getTextContent();
+	}
+
+	/**
+	 * Returns a xml representation as {@link String} object of the given {@link Node} object
+	 * 
+	 * @param node
+	 *            the {@link Node} object
+	 * @param outputProperties
+	 *            the output properties for the as {@link Transformer} object
+	 * @return a xml representation as {@link String} object of the given {@link Node} object
+	 * @throws TransformerException
+	 *             is thrown if an unrecoverable error occurs during the course of the
+	 *             transformation
+	 */
 	public static String toString(Node node, Map<String, String> outputProperties)
 		throws TransformerException
 	{
@@ -71,6 +162,19 @@ public class NodeExtensions
 		return stringWriter.toString();
 	}
 
+	/**
+	 * Creates a new {@link DOMResult} object from the given {@link String} object
+	 * 
+	 * @param xml
+	 *            the xml as {@link String} object
+	 * @return a new {@link DOMResult} object from the given {@link String} object
+	 * @throws ParserConfigurationException
+	 *             if a DocumentBuilder cannot be created which satisfies the configuration
+	 *             requested.
+	 * @throws TransformerException
+	 *             is thrown if an unrecoverable error occurs during the course of the
+	 *             transformation
+	 */
 	public static DOMResult toDOMResult(String xml)
 		throws ParserConfigurationException, TransformerException
 	{
@@ -80,4 +184,5 @@ public class NodeExtensions
 		transformer.transform(streamSource, domResult);
 		return domResult;
 	}
+
 }